requirements.txt
requests>=2.25.1
beautifulsoup4>=4.9.3
pandas>=1.3.0
ollama>=0.1.0
flask>=2.0.1
python-dateutil>=2.8.2
jinja2>=3.0.0
werkzeug>=2.0.0
nltk>=3.6.0
PyPDF2>=2.0.0
numpy>=1.20.0
PyMuPDF>=1.19.0
pdfplumber>=0.7.0
Pillow>=8.0.0
pytesseract>=0.3.8
